% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_required_packages}
\alias{install_required_packages}
\title{Install Required Packages
This function checks if the necessary packages are installed and installs them if not.}
\usage{
install_required_packages()
}
\description{
Install Required Packages
This function checks if the necessary packages are installed and installs them if not.
}
