% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_Explained_Function.R
\name{PVE}
\alias{PVE}
\title{Variance Explained Function
Calculate the percent variation associated with any covariate of interest}
\usage{
PVE(
  counts,
  metadata,
  covariates,
  numeric_covariates,
  cutoff = NULL,
  output_dir,
  rank = 10
)
}
\arguments{
\item{counts}{Matrix of normalized RNA-seq data with genes in rows and samples in columns.}

\item{metadata}{Dataframe of metadata for RNA-seq samples. Samples should be in rows and covariates in columns.}

\item{covariates}{List of covariates in `metadata` to be assessed.}

\item{numeric_covariates}{Character vector of covariates that are numeric; others will be treated as factors.}

\item{cutoff}{Optional threshold for a horizontal line in the variance barplot. Default is NULL.}

\item{output_dir}{Path to save output files.}

\item{rank}{Integer specifying the number of PCs to retain (from PCAtools). Default is 10. Use NULL for all PCs.}
}
\value{
Dataframe with `variable` and `variance_explained` columns; generates a screeplot and a variance barplot.
}
\description{
Variance Explained Function
Calculate the percent variation associated with any covariate of interest
}
\examples{
# Example usage:
PVE(counts = my_counts_matrix, metadata = my_metadata_df,
    covariates = c("Age", "Sex"), numeric_covariates = "Age",
    cutoff = 5, output_dir = "path/to/output", rank = 10)
}
