% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stepwise_Variance_Regression.R
\name{SVR}
\alias{SVR}
\title{Stepwise Variance Regression
Iterative regression to assess covariate variance colinearilty}
\usage{
SVR(data, metadata, covariates, numeric_covariates, output_dir, rank = 10)
}
\arguments{
\item{data}{Matrix of normalized RNA-seq data with genes in rows and samples in columns.}

\item{metadata}{Dataframe of metadata for RNA-seq samples. Samples should be in rows and covariates in columns.}

\item{covariates}{List of covariates in `metadata` to be assessed.}

\item{numeric_covariates}{Character vector of covariates that are numeric; others will be treated as factors.}

\item{output_dir}{Path to save output files.}

\item{rank}{Integer specifying the number of PCs to retain (from PCAtools). Default is 10. Use NULL for all PCs.}
}
\value{
Dataframe of covariates ordered by greatest association with variance after each iterative regression.
}
\description{
Stepwise Variance Regression
Iterative regression to assess covariate variance colinearilty
}
\examples{
# Example usage:
SVR(data = my_counts_matrix, metadata = my_metadata_df,
    covariates = c("Age", "Sex"), numeric_covariates = "Age",
    output_dir = "path/to/output", rank = 10)
}
