% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stepwise_Variance_Regression.R
\name{SVR}
\alias{SVR}
\title{Stepwise Variance Regression
Iterative regression to assess covariate variance colinearilty}
\usage{
SVR(data, metadata, covariates, numeric_covariates, output_dir, rank = 10)
}
\arguments{
\item{metadata}{Dataframe of metadata for RNA-seq samples. Samples should be in rows and covariates in columns.}

\item{covariates}{List of covariates in `metadata` to be assessed.}

\item{numeric_covariates}{Character vector of covariates that are numeric; others will be treated as factors.}

\item{output_dir}{Path to save output files.}

\item{rank}{Integer specifying the number of PCs to retain (from PCAtools). Default is 10. Use NULL for all PCs.}

\item{counts}{Matrix of normalized RNA-seq data with genes in rows and samples in columns.}

\item{cutoff}{Optional threshold for a horizontal line in the variance barplot. Default is NULL.}
}
\value{
Dataframe with `variable` and `variance_explained` columns; generates a screeplot and a variance barplot.
}
\description{
Stepwise Variance Regression
Iterative regression to assess covariate variance colinearilty
}
\examples{
# Example usage:
PVE(counts = my_counts_matrix, metadata = my_metadata_df,
    covariates = c("Age", "Sex"), numeric_covariates = "Age",
    cutoff = 5, output_dir = "path/to/output", rank = 10)
}
